--- project2/OLS-GD.csv
@@ -1,81 +1,0 @@
-MSE,Momentum,Learning-rate,Iterations
-2.708903397607089,0.9,0.01,57.0
-2.821539280190432,0.9,0.01,97.0
-2.698701173250817,0.9,0.01,57.0
-2.7116158476197247,0.9,0.01,57.0
-2.702153571717108,0.9,0.01,57.0
-2.714851061661641,0.9,0.01,57.0
-2.7988195726243226,0.9,0.01,97.0
-2.7157482122160985,0.9,0.01,57.0
-2.6941412286188813,0.9,0.01,57.0
-2.7049869506433732,0.9,0.01,57.0
-2.7926584119304594,0.9,0.01,97.0
-2.676827760296667,0.9,0.01,57.0
-2.750392196906381,0.9,0.01,58.0
-2.7556115297826995,0.9,0.01,57.0
-2.6370218065070157,0.9,0.01,56.0
-2.8053238983229125,0.9,0.01,58.0
-2.694420713895439,0.9,0.01,57.0
-2.6668980672187974,0.9,0.01,57.0
-2.7002426679858607,0.9,0.01,57.0
-2.7526671647657803,0.9,0.01,57.0
-2.573653242892997,0.9,0.01,57.0
-2.5736864657468295,0.9,0.01,57.0
-2.5736740131276217,0.9,0.01,57.0
-2.5736921774669113,0.9,0.01,57.0
-2.5736997554757477,0.9,0.01,25.0
-2.573642461245463,0.9,0.01,57.0
-2.5736856646128814,0.9,0.01,57.0
-2.573716860099627,0.9,0.01,56.0
-2.5736826772490455,0.9,0.01,57.0
-2.573671775145986,0.9,0.01,57.0
-2.5737083211054506,0.9,0.01,57.0
-2.57363198153019,0.9,0.01,57.0
-2.5735324489315126,0.9,0.01,136.0
-2.573689183808466,0.9,0.01,57.0
-2.573663366693377,0.9,0.01,57.0
-2.573707196191275,0.9,0.01,57.0
-2.573612212009656,0.9,0.01,133.0
-2.573684244272035,0.9,0.01,57.0
-2.5736941494343935,0.9,0.01,57.0
-2.5736948109241373,0.9,0.01,57.0
-2.573482267825041,0.0,0.01,802.0
-2.572688394048572,0.0,0.01,793.0
-2.5724081809639214,0.0,0.01,914.0
-2.5716874410972688,0.0,0.01,927.0
-2.573659424246842,0.0,0.01,786.0
-2.572535977342628,0.0,0.01,866.0
-2.5728923478965045,0.0,0.01,929.0
-2.5723762350099655,0.0,0.01,877.0
-2.573207268737523,0.0,0.01,815.0
-2.572315775905661,0.0,0.01,995.0
-2.5735135904656437,0.0,0.01,753.0
-2.5729150567037355,0.0,0.01,850.0
-2.5724242607803034,0.0,0.01,945.0
-2.573731875808239,0.0,0.01,813.0
-2.5731687846039444,0.0,0.01,694.0
-2.5717310369012116,0.0,0.01,940.0
-2.5731670010846934,0.0,0.01,814.0
-2.5732880405579905,0.0,0.01,690.0
-2.573332915993501,0.0,0.01,800.0
-2.573263650200808,0.0,0.01,807.0
-2.5738044247033085,0.0,0.001,5750.0
-2.5737010801914564,0.0,0.001,5949.0
-2.573780647119054,0.0,0.001,6012.0
-2.5738251439561997,0.0,0.001,5662.0
-2.573067198540182,0.0,0.001,6667.0
-2.5737564144898952,0.0,0.001,5990.0
-2.5735279453232893,0.0,0.001,6147.0
-2.5702656570289144,0.0,0.001,7457.0
-2.5734263734357494,0.0,0.001,6101.0
-2.573732145561524,0.0,0.001,6138.0
-2.5732991738278566,0.0,0.001,6351.0
-2.5737393814373455,0.0,0.001,6114.0
-2.5738517377220353,0.0,0.001,5236.0
-2.5732985796217553,0.0,0.001,6573.0
-2.5738589755664334,0.0,0.001,5414.0
-2.5732839554893236,0.0,0.001,6494.0
-2.573624970988958,0.0,0.001,5953.0
-2.573755811543968,0.0,0.001,5873.0
-2.5733435496165034,0.0,0.001,6499.0
-2.5734190788657587,0.0,0.001,6536.0
