--- project2/OLS-SGD.csv
@@ -1,81 +1,0 @@
-MSE,Momentum,Learning-rate,Iterations,Epoch,Batch-size
-3.520092658876546,0.0,0.01,84.0,100.0,5.0
-2.775710015365289,0.0,0.01,106.0,100.0,10.0
-2.7467380736202887,0.9,0.01,103.0,100.0,15.0
-2.8426538320103667,0.9,0.01,24.0,100.0,20.0
-2.791594193373033,0.9,0.01,24.0,100.0,25.0
-2.714642859649941,0.8,0.01,91.0,100.0,30.0
-2.6450789116238944,0.9,0.01,13.0,100.0,35.0
-2.622411868550299,0.9,0.01,41.0,100.0,40.0
-2.6329981911735736,0.8,0.01,96.0,100.0,45.0
-2.574228681544949,0.9,0.01,46.0,100.0,50.0
-2.698388301099048,0.9,0.01,99.0,100.0,55.0
-2.62013215919061,0.9,0.01,93.0,100.0,60.0
-2.643266940726782,0.9,0.01,53.0,100.0,65.0
-2.7378285882440663,0.9,0.01,53.0,100.0,70.0
-2.690230526993177,0.9,0.01,87.0,100.0,75.0
-2.6354110960755053,0.9,0.01,50.0,100.0,80.0
-2.6797683158216854,0.9,0.01,52.0,100.0,85.0
-2.673479650474831,0.9,0.01,54.0,100.0,90.0
-2.741717054796024,0.9,0.01,93.0,100.0,95.0
-2.731359730832119,0.9,0.01,57.0,100.0,100.0
-2.7547858532350724,0.0,0.001,384.0,464.0,5.0
-2.733239083386001,0.0,0.01,410.0,464.0,10.0
-2.7667505776617487,0.0,0.01,202.0,464.0,15.0
-2.7708708589538644,0.0,0.01,225.0,464.0,20.0
-2.810850253498838,0.0,0.01,205.0,464.0,25.0
-2.711857040131538,0.0,0.01,253.0,464.0,30.0
-2.6508029776599966,0.6000000000000001,0.01,161.0,464.0,35.0
-2.6317475319129278,0.1,0.01,316.0,464.0,40.0
-2.601807996461068,0.0,0.01,383.0,464.0,45.0
-2.5652895816603376,0.0,0.01,369.0,464.0,50.0
-2.5441255226105253,0.8,0.01,289.0,464.0,55.0
-2.5823062067515234,0.9,0.01,130.0,464.0,60.0
-2.5815277980912374,0.9,0.01,89.0,464.0,65.0
-2.5984410310462436,0.9,0.01,124.0,464.0,70.0
-2.5900445725164065,0.9,0.01,52.0,464.0,75.0
-2.573843671792755,0.9,0.01,50.0,464.0,80.0
-2.5738161190450533,0.9,0.01,53.0,464.0,85.0
-2.5737817471126005,0.9,0.01,54.0,464.0,90.0
-2.5737502380990613,0.9,0.01,94.0,464.0,95.0
-2.573646669121025,0.9,0.01,57.0,464.0,100.0
-2.8357288311173052,0.0,0.001,1380.0,2154.0,5.0
-2.699118487929238,0.0,0.001,2027.0,2154.0,10.0
-2.7340483060445147,0.0,0.001,1493.0,2154.0,15.0
-2.745408085050285,0.0,0.001,1378.0,2154.0,20.0
-2.7812070472452106,0.0,0.001,1416.0,2154.0,25.0
-2.6881717976843453,0.0,0.01,360.0,2154.0,30.0
-2.6406586042920366,0.0,0.01,311.0,2154.0,35.0
-2.6303323104466894,0.0,0.01,373.0,2154.0,40.0
-2.616480552874532,0.0,0.01,426.0,2154.0,45.0
-2.588806673603569,0.0,0.01,382.0,2154.0,50.0
-2.5535888806741465,0.0,0.01,1102.0,2154.0,55.0
-2.5812867859784676,0.0,0.01,898.0,2154.0,60.0
-2.5815257264595046,0.0,0.01,798.0,2154.0,65.0
-2.598117330690982,0.0,0.01,980.0,2154.0,70.0
-2.5898196244423968,0.0,0.01,780.0,2154.0,75.0
-2.5736029105541935,0.0,0.01,698.0,2154.0,80.0
-2.5731732058148546,0.0,0.01,785.0,2154.0,85.0
-2.5735485585347355,0.0,0.01,783.0,2154.0,90.0
-2.573303094702942,0.0,0.01,823.0,2154.0,95.0
-2.572369183752625,0.0,0.01,931.0,2154.0,100.0
-2.7191443389964634,0.9,1e-05,7555.0,10000.0,5.0
-2.7065732933495203,0.0,0.0001,9581.0,10000.0,10.0
-2.7175119555858833,0.0,0.0001,8016.0,10000.0,15.0
-2.724820697931328,0.0,0.0001,7654.0,10000.0,20.0
-2.8016780764321596,0.0,0.001,1459.0,10000.0,25.0
-2.707110820301012,0.0,0.001,2373.0,10000.0,30.0
-2.649887354663031,0.0,0.001,2612.0,10000.0,35.0
-2.626614770114986,0.0,0.001,2997.0,10000.0,40.0
-2.6127618293673502,0.0,0.001,2685.0,10000.0,45.0
-2.5884962335945634,0.0,0.001,2939.0,10000.0,50.0
-2.5547424299137873,0.0,0.001,6848.0,10000.0,55.0
-2.5817228807206645,0.0,0.001,6639.0,10000.0,60.0
-2.581479080487574,0.0,0.001,6758.0,10000.0,65.0
-2.598401557510108,0.0,0.001,6886.0,10000.0,70.0
-2.589956145717627,0.0,0.001,6429.0,10000.0,75.0
-2.573603714162179,0.0,0.001,6133.0,10000.0,80.0
-2.573597355722392,0.0,0.001,6046.0,10000.0,85.0
-2.573536082308976,0.0,0.001,6313.0,10000.0,90.0
-2.5737372012758257,0.0,0.001,6112.0,10000.0,95.0
-2.57377894450545,0.0,0.001,5702.0,10000.0,100.0
